Update has_side_effects decorator to support function views
I have added a kwarg to the decorator that can be used to manage
whether the side effects run based on the return value of the function.
This can be used for example when decorating a view function that will
return an HttpResponse object even if the function logically failed.

The default value for the kwarg is a function called
`http_response_check` which will return False if the inner function
returns an HttpResponse object with a status code of 4xx, 5xx, so the
decorator can be used out of the box to decorate view functions.