Ensure that has_side_effects runs on transaction.on_commit (#24)
This PR ensures that when a function is decorated with the
has_side_effects decorator that the call to run_side_effects is deferred
using the transaction.on_commit function so that it will only run once
the transaction has committed. (The transaction.atomic decorator will
only create a new transaction if one does not already exist - otherwise
it uses the existing transaction to create a savepoint.) This should
prevent a class of errors that occurs when the side-effect is running in
different process (e.g. as a job on a queue) but relies on data that the
source function is creating within a transaction. In this instance the
side-effect would only be invoked once the transaction has been
committed, so it can safely assume that the data has been written to the
database.