Fix issue with disable_side_effects
When running inside an existing transaction the disable_side_effects
decorator has a bug. It disables and then enables the registry so
that side-effects don't run, but if the call to `run_side_effects`
is deferred using `transaction.on_commit` it may be called after
the context manager has exited, in which case the registry will
have been re-enabled. In this case the side-effects will run.

The fix is to pull some of the logic around suppression out so that
it runs immediately on calling `run_side_effects` - if the registry
is disabled at that point it will fire the signal and return
immediately.